Extract CactusRef to its own module
Rename to Rc
